Face3: Remove from core.
