--- docs/api/en/core/Face3.html
@@ -1,100 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-	<head>
-		<meta charset="utf-8" />
-		<base href="../../../" />
-		<script src="page.js"></script>
-		<link type="text/css" rel="stylesheet" href="page.css" />
-	</head>
-	<body>
-		<h1>[name]</h1>
-
-		<p class="desc">
-		Triangular face used in deprecated geometry. These are created automatically for all
-		standard geometry types, however if you are building a custom geometry you will have to
-		create them manually.
-		</p>
-
-		<h2>Constructor</h2>
-
-		<h3>[name]( [param:Integer a], [param:Integer b], [param:Integer c], [param:Vector3 normal], [param:Color color], [param:Integer materialIndex] )</h3>
-		<p>
-		a — Vertex A index.<br />
-		b — Vertex B index.<br />
-		c — Vertex C index.<br /><br />
-
-		normal — (optional) Face normal ([page:Vector3 Vector3]) or array of vertex normals.
-		If a single vector is passed in, this sets [page:.normal], otherwise if an array of three
-		vectors is passed in this sets [page:.vertexNormals]<br /><br />
-
-		color — (optional) Face [page:Color color] or array of vertex [page:Color colors].
-		If a single vector is passed in, this sets [page:.color], otherwise if an array of three
-		vectors is passed in this sets [page:.vertexColors]<br /><br />
-
-		materialIndex — (optional) which index of an array of materials to associate
-		with the face.
-		</p>
-
-		<h2>Properties</h2>
-
-		<h3>[property:Integer a]</h3>
-		<p>
-		Vertex A index.
-		</p>
-
-		<h3>[property:Integer b]</h3>
-		<p>
-		Vertex B index.
-		</p>
-
-		<h3>[property:Integer c]</h3>
-		<p>
-		Vertex C index.
-		</p>
-
-		<h3>[property:Vector3 normal]</h3>
-		<p>
-		Face normal - vector showing the direction of the Face3. If calculated automatically,
-		this is the normalized cross product of two edges of the
-		triangle. Default is *(0, 0, 0)*.
-		</p>
-
-		<h3>[property:Color color]</h3>
-		<p>
-		Face color - for this to be used a material's [page:Material.vertexColors vertexColors] property
-		must be set to *true*.
-		</p>
-
-		<h3>[property:Array vertexNormals]</h3>
-		<p>
-		Array of 3 [page:Vector3 vertex normals].
-		</p>
-
-		<h3>[property:Array vertexColors]</h3>
-		<p>
-		Array of 3 vertex colors - for these to be used a material's [page:Material.vertexColors vertexColors] property
-		must be set to *true*.
-		</p>
-
-
-		<h3>[property:Integer materialIndex]</h3>
-		<p>
-		Material index (points to an index in the associated array of materials). Default is *0*.
-		</p>
-
-		<h2>Methods</h2>
-
-		<h3>[method:Face3 clone]()</h3>
-		<p>Creates a new clone of the Face3 object.</p>
-
-		<h3>[method:Face3 copy]( [param:Face3 face3] )</h3>
-		<p>Copy the parameters of another Face3 into this.</p>
-
-
-		<h2>Source</h2>
-
-		<p>
-			[link:https://github.com/mrdoob/three.js/blob/master/src/[path].js src/[path].js]
-		</p>
-	</body>
-</html>
