+++ docs/api/zh/core/Raycaster.html
@@ -175,7 +175,7 @@
 		<p>
 			[page:Float distance] —— 射线投射原点和相交部分之间的距离。<br />
 			[page:Vector3 point] —— 相交部分的点（世界坐标）<br />
-			[page:Face3 face] —— 相交的面<br />
+			[page:Object face] —— 相交的面<br />
 			[page:Integer faceIndex] —— 相交的面的索引<br />
 			[page:Object3D object] —— 相交的物体<br />
 			[page:Vector2 uv] —— 相交部分的点的UV坐标。<br />
