+++ examples/jsm/deprecated/Geometry.d.ts
@@ -1,7 +1,6 @@
 import {
 	Vector3,
 	Color,
-	Face3,
 	Vector2,
 	Vector4,
 	Box3,
@@ -17,7 +16,7 @@ import {
 } from '../../../src/Three';
 
 /**
- * @deprecated Use {@link Face3} instead.
+ * @deprecated Use Face3 instead.
  */
 
 export interface MorphTarget {
@@ -37,8 +36,6 @@ export interface MorphNormals {
 
 /**
  * Base class for geometries
- *
- * see {@link https://github.com/mrdoob/three.js/blob/master/src/core/Geometry.js|src/core/Geometry.js}
  */
 export class Geometry extends EventDispatcher {
 
@@ -277,5 +274,101 @@ export class Geometry extends EventDispatcher {
 	bones: Bone[];
 	animation: AnimationClip;
 	animations: AnimationClip[];
+
+}
+
+/**
+ * Triangle face.
+ */
+export class Face3 {
+
+	/**
+	 * @param a Vertex A index.
+	 * @param b Vertex B index.
+	 * @param c Vertex C index.
+	 * @param normal Face normal or array of vertex normals.
+	 * @param color Face color or array of vertex colors.
+	 * @param materialIndex Material index.
+	 */
+	constructor(
+		a: number,
+		b: number,
+		c: number,
+		normal?: Vector3,
+		color?: Color,
+		materialIndex?: number
+	);
+	constructor(
+		a: number,
+		b: number,
+		c: number,
+		normal?: Vector3,
+		vertexColors?: Color[],
+		materialIndex?: number
+	);
+	constructor(
+		a: number,
+		b: number,
+		c: number,
+		vertexNormals?: Vector3[],
+		color?: Color,
+		materialIndex?: number
+	);
+	constructor(
+		a: number,
+		b: number,
+		c: number,
+		vertexNormals?: Vector3[],
+		vertexColors?: Color[],
+		materialIndex?: number
+	);
+
+	/**
+	 * Vertex A index.
+	 */
+	a: number;
+
+	/**
+	 * Vertex B index.
+	 */
+	b: number;
+
+	/**
+	 * Vertex C index.
+	 */
+	c: number;
+
+	/**
+	 * Face normal.
+	 * @default new THREE.Vector3()
+	 */
+	normal: Vector3;
+
+	/**
+	 * Array of 3 vertex normals.
+	 * @default []
+	 */
+	vertexNormals: Vector3[];
+
+	/**
+	 * Face color.
+	 * @default new THREE.Color()
+	 */
+	color: Color;
+
+	/**
+	 * Array of 3 vertex colors.
+	 * @default []
+	 */
+	vertexColors: Color[];
+
+	/**
+	 * Material index (points to {@link Mesh.material}).
+	 * @default 0
+	 */
+	materialIndex: number;
+
+	clone(): Face3;
+	copy( source: Face3 ): this;
 
 }
